Shmem: refactor, clean includes, make header only
